dependencies(0.1.1): Update KUtils to 0.11.2
