+++ pom.xml
@@ -6,13 +6,13 @@
 
     <groupId>com.supergrecko.questionbot</groupId>
     <artifactId>QuestionBot</artifactId>
-    <version>0.0.0</version>
+    <version>0.1.1</version>
 
     <properties>
         <project.repository>https://github.com/supergrecko/QuestionBot</project.repository>
         <jvm.target>1.8</jvm.target>
         <kotlin.version>1.3.50</kotlin.version>
-        <kutils.version>0.11.1</kutils.version>
+        <kutils.version>0.11.2</kutils.version>
         <slf4j.version>1.7.28</slf4j.version>
     </properties>
