+++ src/main/kotlin/com/supergrecko/questionbot/commands/UtilityCommands.kt
@@ -18,7 +18,6 @@ fun utilityCommands() = commands {
     }
 
     command("Uptime") {
-        requiresGuild = true
         description = "Displays how long the bot has been running."
         execute {
             val milliseconds = Date().time - startTime.time
